+++ uamqp/async/authentication_async.py
@@ -50,7 +50,7 @@ class CBSAsyncAuthMixin(authentication.CBSAuthMixin):
         in_progress = False
         auth_status = await self.loop.run_in_executor(None, functools.partial(self._cbs_auth.get_status))
         auth_status = constants.CBSAuthStatus(auth_status)
-        if auth_status == constants.CBSAuthStatus.Failure:
+        if auth_status == constants.CBSAuthStatus.Error:
             if self.retries >= self._retry_policy.retries:
                 _logger.warning("Authentication Put-Token failed. Retries exhausted.")
                 raise errors.TokenAuthFailure(*self._cbs_auth.get_failure_info())
@@ -60,6 +60,8 @@ class CBSAsyncAuthMixin(authentication.CBSAuthMixin):
                 await asyncio.sleep(self._retry_policy.backoff)
                 await self.loop.run_in_executor(None, functools.partial(self._cbs_auth.authenticate))
                 in_progress = True
+        elif auth_status == constants.CBSAuthStatus.Failure:
+            errors.AuthenticationException("Failed to open CBS authentication link.")
         elif auth_status == constants.CBSAuthStatus.Expired:
             raise errors.TokenExpired("CBS Authentication Expired.")
         elif auth_status == constants.CBSAuthStatus.Timeout: